+++ src/features/RecipeDisplay/components/IngredientDirectionsRow.tsx
@@ -7,10 +7,10 @@ import {
     Typography,
 } from "@mui/material";
 import React from "react";
-import Directions from "../common/Directions";
-import IngredientItem from "../IngredientItem";
-import { Recipe } from "../../global/types/types";
-import { useScaleOptions } from "../../util/ScalingContext";
+import Directions from "views/common/Directions";
+import IngredientItem from "views/IngredientItem";
+import { Recipe } from "global/types/types";
+import { useScaleOptions } from "util/ScalingContext";
 
 interface Props {
     recipe: Recipe,
